requirements.txt
streamlit>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
pyttsx3>=2.90
Pillow>=10.0.0
qrcode[pil]>=7.3.1
